Changed add_entities and fixed directory bug
Changed the add_entities method so that it can also add entities from
an existing instance of the DXFGeometry class.

Also fixed a bug in my modifications to Kerry’s code where looking for
a DXF file in the current directory without using the ‘.’ to denote the
current directory would make the current directory ‘’ which wouldn’t
work.